+++ AngularClient/src/main/core/rxjs-extensions.ts
@@ -2,3 +2,4 @@
 import "rxjs/add/operator/catch";
 import "rxjs/add/operator/finally";
 import "rxjs/add/operator/mergeMap";
+import "rxjs/add/operator/share";