+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 ### Changelog
 
+**0.1.2**
+
+* NumericValue -> DecimalValue - coni2k
+* AngularClient - AppErrorHandler sourceMap fix - coni2k
+
 **0.1.1**
 
-* Social Logins were removed
+* Social Logins were removed - coni2k
 
 **0.1.0**
 
-* Initial commit
+* Initial commit - coni2k